20/03/2019 23/51

+++ src/main/java/ua/in/bibi/ecommerceonlineshopping/controller/BrandsController.java
@@ -6,11 +6,10 @@ import org.springframework.web.bind.annotation.*;
 import ua.in.bibi.ecommerceonlineshopping.dto.request.BrandsRequest;
 import ua.in.bibi.ecommerceonlineshopping.dto.response.BrandsResponse;
 import ua.in.bibi.ecommerceonlineshopping.dto.response.DataResponse;
-import ua.in.bibi.ecommerceonlineshopping.entity.Brands;
 import ua.in.bibi.ecommerceonlineshopping.exception.WrongInputException;
-import ua.in.bibi.ecommerceonlineshopping.repository.BrandsRepository;
 import ua.in.bibi.ecommerceonlineshopping.service.BrandsService;
 
+import javax.validation.Valid;
 import java.util.List;
 
 @CrossOrigin
@@ -21,52 +20,72 @@ public class BrandsController {
     @Autowired
     private BrandsService brandsService;
 
-    @Autowired
-    private BrandsRepository brandsRepository;
 
 //    create
 //    PostMapping("/brand")
 @PostMapping
-public BrandsResponse create(@RequestBody BrandsRequest brandsRequest) {
+public BrandsResponse create(@RequestBody @Valid BrandsRequest brandsRequest) {
     return brandsService.create(brandsRequest);
     }
 
 //    read
-//    @GetMapping("/brands")
 //    @GetMapping
-//    public List<BrandsResponse> findAll() {
+@GetMapping("/all")
+public List<BrandsResponse> findAll() {
 //        System.out.println("GET ALL BRANDS");
-//        return brandsService.findAll();
-//    }
+    return brandsService.findAll();
+}
 
     @GetMapping
-    public DataResponse<BrandsResponse> getPersons(@RequestParam(required = false) String value,
+    public DataResponse<BrandsResponse> getBrandsSortPaginationFindValue(
+            @RequestParam(required = false) String value,
                                                    @RequestParam Integer page,
                                                    @RequestParam Integer size,
                                                    @RequestParam String sortFieldName,
                                                    @RequestParam Sort.Direction direction) {
-        System.out.println("GET ALL PERSONS");
-        return brandsService.findAll(value, page, size, sortFieldName, direction);
+        if (page >= 0) {
+            if (size > 0) {
+//                System.out.println("GET ALL BRANDS pageble");
+                return brandsService.findAll(value, page, size, sortFieldName, direction);
+            } else {
+                throw new IllegalArgumentException("SIZE id must not be less than ONE!");
+            }
+        } else {
+            throw new IllegalArgumentException("PAGE id must not be less than ZERO!");
+        }
+
     }
 
-//    @GetMapping("/{id}")
-//    public BrandsResponse getBrandById(@PathVariable Long id) {
-//        System.out.println("Get brand by id : " + id);
-//        return brandsService.findOne(id);
-//    }
+    @GetMapping("/{id}")
+    public BrandsResponse getBrandById(@PathVariable Long id) {
+        if (id > 0) {
+//            System.out.println("Get brand by id : " + id);
+            return brandsService.findOneById(id);
+        } else {
+            throw new IllegalArgumentException("BRAND id must not be less than ONE!");
+        }
+    }
 
 //    update
 //    @PutMapping("/brand")
 @PutMapping
-public BrandsResponse update(@RequestBody BrandsRequest brandsRequest, @RequestParam Long id) throws WrongInputException {
-    return brandsService.update(brandsRequest, id);
+public BrandsResponse update(@RequestBody @Valid BrandsRequest brandsRequest, @RequestParam Long id) throws WrongInputException {
+    if (id > 0) {
+        return brandsService.update(brandsRequest, id);
+    } else {
+        throw new IllegalArgumentException("BRAND id must not be less than ONE!");
+    }
     }
 
 //    delete
 //    @DeleteMapping("/brand")
 @DeleteMapping
     public void delete(@RequestParam Long id) throws WrongInputException {
-    brandsService.delete(id);
+    if (id > 0) {
+        brandsService.delete(id);
+    } else {
+        throw new IllegalArgumentException("BRAND id must not be less than ONE!");
+    }
     }
